+++ Clases/Clinica_dental_PI/src/main/java/com/example/clinica_dental_pi/service/TurnoService.java
@@ -1,7 +1,11 @@
 package com.example.clinica_dental_pi.service;
 
 
+import com.example.clinica_dental_pi.Repository.OdontologoRepository;
+import com.example.clinica_dental_pi.Repository.PacienteRepository;
 import com.example.clinica_dental_pi.Repository.TurnoRepository;
+import com.example.clinica_dental_pi.exceptions.BadRequestException;
+import com.example.clinica_dental_pi.exceptions.ResourceNotFoundException;
 import com.example.clinica_dental_pi.model.Turno;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,6 +18,10 @@ public class TurnoService {
 
     @Autowired
     private TurnoRepository turnoRepository;
+    @Autowired
+    private PacienteRepository pacienteRepository;
+    @Autowired
+    private OdontologoRepository odontologoRepository;
 
     public TurnoService(TurnoRepository turnoRepository) {
         this.turnoRepository = turnoRepository;
@@ -23,11 +31,19 @@ public class TurnoService {
         return turnoRepository.findAll();
     }
 
-    public Turno registrarTurno(Turno turno){
+    public Turno registrarTurno(Turno turno) throws BadRequestException {
+        if(pacienteRepository.findById(turno.getPaciente().getId()) == null ){
+            throw new BadRequestException("No existe ningun paciente con ese id: "+ turno.getPaciente().getId());
+        } else if (odontologoRepository.findById(turno.getOdontologo().getId())== null){
+            throw new BadRequestException("No existe ningun odontologo con ese id: "+ turno.getOdontologo().getId());
+        }
         return turnoRepository.save(turno);
     }
 
-    public void eliminarTurno(Integer id) {
+    public void eliminarTurno(Integer id) throws ResourceNotFoundException {
+        if(buscar(id)==null){
+            throw new ResourceNotFoundException("No existe ningun turno con ese id: " +id);
+        }
         turnoRepository.deleteById(id);
     }
